refactor: split conversation API routes by domain
