+++ src/agent_manager/api/deps.py
@@ -4,8 +4,9 @@ from __future__ import annotations
 
 import logging
 from dataclasses import dataclass
+from typing import Annotated
 
-from fastapi import HTTPException, Request
+from fastapi import Depends, HTTPException, Request
 
 from agent_manager.application import ConversationService
 from agent_manager.domain import Principal
@@ -52,6 +53,11 @@ def get_principal(request: Request) -> Principal:
     except TokenError as exc:
         logger.warning("token verification failed: %s", exc)
         raise HTTPException(status_code=401, detail=str(exc)) from None
+
+
+Service = Annotated[ConversationService, Depends(get_service)]
+Caller = Annotated[Principal, Depends(get_principal)]
+Identity = Annotated[CallerIdentity, Depends(get_caller_identity)]
 
 
 def _select_token(request: Request, identity: CallerIdentity) -> str | None:
