+++ src/agent_manager/api/routes/__init__.py
@@ -1,0 +1,5 @@
+"""Conversation API router assembled from domain routers."""
+
+from agent_manager.api.routes.router import router
+
+__all__ = ["router"]
